đau lưng quáaaa

+++ src/main/java/hcmute/lp/backend/repository/ProductRepository.java
@@ -6,6 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
     boolean existsByName(String name);
+
 }
